+++ src/main/java/com/api/product/controller/SaleController.java
@@ -1,22 +1,17 @@
 package com.api.product.controller;
 
+import com.api.product.model.Sale;
+import com.api.product.repository.SalesRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.*;
+
 import java.math.BigDecimal;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.api.product.model.Sale;
-import com.api.product.repository.SalesRepository;
 
 @RestController
 public class SaleController {
@@ -35,7 +30,7 @@ public class SaleController {
 	}
 	
 	@RequestMapping(value = "/sale/{id}", method = RequestMethod.GET)
-	public Sale findById(@PathVariable BigDecimal id) {
+	public Sale findById(@PathVariable Long id) {
 		return saleRepository.findOne(id);
 	}
 	
@@ -54,9 +49,7 @@ public class SaleController {
 	}
 	
 	@RequestMapping(value = "/sale/delete/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable BigDecimal id) {
+	public void delete(@PathVariable Long id) {
 		this.saleRepository.delete(id);
 	}
-	
-	
 }
