+++ src/main/java/com/api/product/repository/CategoryRepository.java
@@ -4,8 +4,6 @@ import com.api.product.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
-
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, BigDecimal> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
